Enable imx8 kernel via configuration layout
Redefenition of makeModulesClosure resolves build blocker and, also,
allow to get rid from default set of kernel modules check adding
own set of modules presence of which need to be verified.

It definetily not the best way of doing - proper way is to define
several hardware profiles with possibility to be chosen from config
layout.

Signed-off-by: Grigoriy Romanov <grigoriy.romanov@unikie.com>