Make images for nxp use kvms optionally
There is implementation of the way how to control presense
of the kvms in a result system. Config file the same for
every image description file and the last ones tune the kvms
package as needed. The boolean flag which control the logic
of kvms presence passed as an attribute from pkgs where it
appears from overlay described in config file.

Can be concluded that attribute sets for kvms-bin and kvms
package going to be built from sources should be the same.

Also: Remove separate image description for kvms in qxp folder

Signed-off-by: Grigoriy Romanov <grigoriy.romanov@unikie.com>